requirements.txt
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
lightning==2.0
xformers
hydra-core
hydra-colorlog
wandb
matplotlib
einops
tqdm
safetensors
scipy
scikit-learn
opencv-python
h5py
rich
digit-interface
